Cached Dao should only cache immutable objects
It's better if caches only store immutable objects, then we won't have
to worry about dealing with invalid/dirty objects when the database
items are modified.

A new VersionedResourceDao interface is created which adds a new
get(resourceId, dataVersion) method, since the resourceId/dataVersion
pair is enough to identify an immutable resource object in the database.